(1446) Handle versioning of serialization
