--- packages/api/server/models/shantytownModel/_common/livingConditions/getLivingConditionsStatuses.js
@@ -1,7 +1,0 @@
-const getStatusesV1 = require('./v1/statuses/main');
-
-const STRATEGIES = {
-    1: getStatusesV1,
-};
-
-module.exports = town => STRATEGIES[1](town);
